+++ src/main/java/ru/yandex/practicum/filmorate/storage/director/DirectorDbStorage.java
@@ -28,7 +28,7 @@ public class DirectorDbStorage implements DirectorStorage {
             ps.setString(1, director.getName());
             return ps;
         }, keyHolder);
-        director.setId(Objects.requireNonNull(keyHolder.getKey()).intValue());
+        director.setId(Objects.requireNonNull(keyHolder.getKey()).longValue());
         return director;
     }
 
@@ -39,7 +39,7 @@ public class DirectorDbStorage implements DirectorStorage {
     }
 
     @Override
-    public Optional<Director> get(int directorId) {
+    public Optional<Director> get(Long directorId) {
         final String sqlQuery = "SELECT DIRECTOR_ID, DIRECTOR_NAME FROM directors WHERE DIRECTOR_ID = ?";
         final List<Director> directors = jdbcTemplate.query(sqlQuery, RowMapper::mapRowToDirector, directorId);
 
@@ -54,7 +54,7 @@ public class DirectorDbStorage implements DirectorStorage {
     }
 
     @Override
-    public boolean delete(int directorId) {
+    public boolean delete(Long directorId) {
         final String sqlQuery = "DELETE FROM directors WHERE DIRECTOR_ID = ?";
         return jdbcTemplate.update(sqlQuery, directorId) > 0;
     }
@@ -62,7 +62,7 @@ public class DirectorDbStorage implements DirectorStorage {
     @Override
     public void setDirectors(List<Film> films) {
         final String inSql = String.join(",", Collections.nCopies(films.size(), "?"));
-        final Map<Integer, Film> filmById = films.stream().collect(Collectors.toMap(Film::getId, (f) -> f));
+        final Map<Long, Film> filmById = films.stream().collect(Collectors.toMap(Film::getId, (f) -> f));
         if(inSql.isEmpty()) return;
 
         jdbcTemplate.query(
@@ -71,10 +71,10 @@ public class DirectorDbStorage implements DirectorStorage {
                         "LEFT JOIN directors AS d ON fd.DIRECTOR_ID = d.DIRECTOR_ID " +
                         "WHERE fd.FILM_ID IN (%s)", inSql),
                 (rs) -> {
-                    final Film film = filmById.get(rs.getInt("FILM_ID"));
+                    final Film film = filmById.get(rs.getLong("FILM_ID"));
 
                     if (null != film.getDirectors()) {
-                        film.addDirector(new Director(rs.getInt("DIRECTOR_ID"), rs.getString("DIRECTOR_NAME")));
+                        film.addDirector(new Director(rs.getLong("DIRECTOR_ID"), rs.getString("DIRECTOR_NAME")));
                     }
                 },
                 films.stream().map(Film::getId).toArray());
