+++ src/main/java/ru/yandex/practicum/filmorate/storage/friends/FriendsDbStorage.java
@@ -14,19 +14,19 @@ public class FriendsDbStorage implements FriendsStorage {
     private final JdbcTemplate jdbcTemplate;
 
     @Override
-    public void addFriend(int userId, int friendId) {
+    public void addFriend(Long userId, Long friendId) {
         String sqlQuery = "INSERT INTO friends (USER_ID, FRIEND_ID) VALUES (?, ?)";
         jdbcTemplate.update(sqlQuery, userId, friendId);
     }
 
     @Override
-    public void deleteFriend(int userId, int friendId) {
+    public void deleteFriend(Long userId, Long friendId) {
         String sqlQuery = "DELETE FROM friends WHERE USER_ID = ? AND FRIEND_ID = ?";
         jdbcTemplate.update(sqlQuery, userId, friendId);
     }
 
     @Override
-    public List<User> commonFriends(int userId, int friendId) {
+    public List<User> commonFriends(Long userId, Long friendId) {
         String sqlQuery = "SELECT  us.USER_ID, EMAIL, LOGIN, USER_NAME, BIRTHDAY " +
                 "FROM users AS us " +
                 "JOIN friends AS u1 ON us.USER_ID = u1.FRIEND_ID " +
@@ -36,7 +36,7 @@ public class FriendsDbStorage implements FriendsStorage {
     }
 
     @Override
-    public List<User> getFriends(int userId) {
+    public List<User> getFriends(Long userId) {
         String sqlQuery = "SELECT u.USER_ID, f.FRIEND_ID, u.EMAIL, u.LOGIN, u.USER_NAME, u.BIRTHDAY " +
                 "FROM friends AS f " +
                 "LEFT JOIN users AS u ON f.FRIEND_ID = u.USER_ID " +
