+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserDbStorage.java
@@ -36,7 +36,7 @@ public class UserDbStorage implements UserStorage {
             else statement.setDate(4, Date.valueOf(birthday));
             return statement;
         }, keyHolder);
-        user.setId(Objects.requireNonNull(keyHolder.getKey()).intValue());
+        user.setId(Objects.requireNonNull(keyHolder.getKey()).longValue());
         return user;
     }
 
@@ -54,7 +54,7 @@ public class UserDbStorage implements UserStorage {
     }
 
     @Override
-    public Optional<User> get(int userId) {
+    public Optional<User> get(Long userId) {
         final String sqlQuery = "SELECT USER_ID, EMAIL, LOGIN, USER_NAME, BIRTHDAY FROM users WHERE USER_ID = ?";
         final List<User> users = jdbcTemplate.query(sqlQuery, RowMapper::mapRowToUser, userId);
         return users.isEmpty() ? Optional.empty() : Optional.of(users.get(0));
@@ -71,7 +71,7 @@ public class UserDbStorage implements UserStorage {
     }
 
     @Override
-    public boolean delete(int userId) {
+    public boolean delete(Long userId) {
         final String sqlQuery = "DELETE FROM users WHERE USER_ID = ?";
         return jdbcTemplate.update(sqlQuery, userId) > 0;
     }
